Better (not perfect) error management for encode transaction API endpoint
