+++ src/api/adamik/chains.ts
@@ -7,7 +7,7 @@ type GetChainsResponse = {
   chains: Record<string, Chain>;
 };
 
-// TODO Better API error management, consistent for for all endpoints
+// TODO Better API error management, consistent for all endpoints
 export const getChains = async (): Promise<Record<string, Chain> | null> => {
   const response = await fetch(`${ADAMIK_API_URL}/chains`, {
     headers: {
