+++ src/api/adamik/encode.ts
@@ -3,10 +3,10 @@
 import { env, ADAMIK_API_URL } from "~/env";
 import { Transaction, PlainTransaction } from "~/utils/types";
 
-// TODO Better API error management, consistent for for all endpoints
+// TODO Better API error management, consistent for all endpoints
 export const transactionEncode = async (
   plainTransaction: PlainTransaction
-): Promise<Transaction | null> => {
+): Promise<Transaction> => {
   const response = await fetch(`${ADAMIK_API_URL}/transaction/encode`, {
     headers: {
       Authorization: env.ADAMIK_API_KEY,
@@ -16,11 +16,17 @@ export const transactionEncode = async (
     body: JSON.stringify({ transaction: { plain: plainTransaction } }),
   });
 
-  if (response.status === 200) {
-    const { transaction }: { transaction: Transaction } = await response.json();
-    return transaction;
-  } else {
-    console.error("encode - backend error");
-    return null;
+  const {
+    transaction,
+    message,
+  }: { transaction: Transaction; message: string } = await response.json();
+
+  const messageString = message && JSON.stringify(message);
+
+  if (messageString) {
+    console.error("encode - backend error:", messageString);
+    throw new Error(messageString);
   }
+
+  return transaction;
 };
