+++ src/api/adamik/staking.ts
@@ -50,7 +50,7 @@ export type DataAddressStateStaking = {
   balances: Balances;
 };
 
-// TODO Better API error management, consistent for for all endpoints
+// TODO Better API error management, consistent for all endpoints
 export const getStaking = (): DataAddressStateStaking => {
   return {
     chainId: "cosmoshub",
