+++ src/api/adamik/validators.ts
@@ -12,7 +12,7 @@ export type ValidatorResponse = {
   }[];
 };
 
-// TODO Better API error management, consistent for for all endpoints
+// TODO Better API error management, consistent for all endpoints
 export const getValidators = async (
   chainId: string,
   offset?: number,
@@ -41,9 +41,11 @@ export const getValidators = async (
     body: JSON.stringify(body),
   });
 
+  const result = await response.json();
+
   if (response.status !== 200) {
-    console.error("validators - backend error:", response.statusText);
+    console.error("validators - backend error:", result);
   }
 
-  return response.json();
+  return result;
 };
