+++ src/app/stake/TransactionForm.tsx
@@ -25,28 +25,38 @@ import { useTransaction } from "~/hooks/useTransaction";
 import { useEncodeTransaction } from "~/hooks/useEncodeTransaction";
 import { amountToSmallestUnit } from "~/utils/helper";
 import { TransactionFormInput, transactionFormSchema } from "~/utils/schema";
-import { Asset, PlainTransaction, TransactionMode } from "~/utils/types";
-import { AssetsSelector } from "./AssetsSelector";
-import { TransactionLoading } from "./TransactionLoading";
+import {
+  Asset,
+  PlainTransaction,
+  TransactionMode,
+  Validator,
+} from "~/utils/types";
+import { TransactionLoading } from "../portfolio/TransactionLoading";
+import { ValidatorSelector } from "./ValidatorSelector";
+import { AssetsSelector } from "../portfolio/AssetsSelector";
 
 type TransactionProps = {
   onNextStep: () => void;
   assets: Asset[];
+  validators: Validator[];
 };
 
-// FIXME Some duplicate logic to put in common with src/app/stake/TransactionModal.tsx
+// FIXME Some duplicate logic to put in common with src/app/portfolio/TransactionForm.tsx
 
-export function TransactionModal({ onNextStep, assets }: TransactionProps) {
+export function TransactionForm({
+  onNextStep,
+  validators,
+  assets,
+}: TransactionProps) {
   const { mutate, isPending, isSuccess } = useEncodeTransaction();
   const form = useForm<TransactionFormInput>({
     resolver: zodResolver(transactionFormSchema),
     defaultValues: {
-      mode: TransactionMode.TRANSFER,
+      mode: TransactionMode.DELEGATE,
       chainId: "",
-      tokenId: "",
       senders: "",
-      recipients: "",
-      amount: 0,
+      validatorAddress: "",
+      amount: undefined,
       useMaxAmount: false,
     },
   });
@@ -57,17 +67,21 @@ export function TransactionModal({ onNextStep, assets }: TransactionProps) {
   const onSubmit = useCallback(
     (formInput: TransactionFormInput) => {
       const plainTransaction: PlainTransaction = {
-        mode: formInput.mode,
+        mode: TransactionMode.DELEGATE,
         chainId: formInput.chainId,
-        tokenId: formInput.tokenId,
-        recipients: formInput.recipients ? [formInput.recipients] : [],
         senders: [formInput.senders],
-        amount: formInput.useMaxAmount
-          ? ""
-          : amountToSmallestUnit(formInput.amount.toString(), decimals),
+        recipients: [],
+        validatorAddress: formInput.validatorAddress ?? "",
         useMaxAmount: formInput.useMaxAmount,
         format: "json", // FIXME Not always the default, should come from chains config
       };
+
+      if (formInput.amount && !formInput.useMaxAmount) {
+        plainTransaction.amount = amountToSmallestUnit(
+          formInput.amount.toString(),
+          decimals
+        );
+      }
 
       // FIXME Hack to be able to provide the pubKey, probably better to refacto
       const pubKey = assets.find(
@@ -81,8 +95,7 @@ export function TransactionModal({ onNextStep, assets }: TransactionProps) {
       }
 
       mutate(plainTransaction, {
-        onSettled: (settledTransaction) => {
-          setTransaction(undefined);
+        onSuccess: (settledTransaction) => {
           setTransactionHash(undefined);
           if (settledTransaction) {
             if (
@@ -90,17 +103,18 @@ export function TransactionModal({ onNextStep, assets }: TransactionProps) {
               settledTransaction.status.errors.length > 0
             ) {
               setErrors(settledTransaction.status.errors[0].message);
+              setTransaction(undefined);
             } else {
               setTransaction(settledTransaction);
             }
           } else {
+            setTransaction(undefined);
             setErrors("API ERROR - Please try again later");
           }
         },
         onError: (error) => {
-          setTransaction(undefined);
           setTransactionHash(undefined);
-          setErrors("API ERROR - Please try again later :" + error.message);
+          setErrors(error.message);
         },
       });
     },
@@ -111,7 +125,7 @@ export function TransactionModal({ onNextStep, assets }: TransactionProps) {
     return <TransactionLoading />;
   }
 
-  if (isSuccess && transaction) {
+  if (isSuccess && transaction?.encoded) {
     return (
       <>
         <h1 className="font-bold text-xl text-center">
@@ -133,7 +147,7 @@ export function TransactionModal({ onNextStep, assets }: TransactionProps) {
           <CollapsibleContent>
             <Textarea
               readOnly
-              value={JSON.stringify(transaction)}
+              value={JSON.stringify(transaction.encoded)}
               className="h-32 text-xs text-gray-500 mt-4"
             />
           </CollapsibleContent>
@@ -144,7 +158,7 @@ export function TransactionModal({ onNextStep, assets }: TransactionProps) {
 
   return (
     <>
-      <h1 className="font-bold text-xl text-center">Transfer</h1>
+      <h1 className="font-bold text-xl text-center">Stake</h1>
       <Form {...form}>
         <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8 px-4">
           <FormField
@@ -165,13 +179,8 @@ export function TransactionModal({ onNextStep, assets }: TransactionProps) {
                       form.setValue("assetIndex", index);
                       form.setValue("chainId", asset.chainId);
                       form.setValue("senders", asset.address);
-                      if (asset.isToken) {
-                        form.setValue("mode", TransactionMode.TRANSFER_TOKEN);
-                        form.setValue(
-                          "tokenId",
-                          asset.contractAddress || asset.assetId
-                        );
-                      }
+                      form.resetField("validatorIndex");
+                      form.resetField("validatorAddress");
                       setDecimals(asset.decimals);
                     }}
                     {...field}
@@ -198,14 +207,37 @@ export function TransactionModal({ onNextStep, assets }: TransactionProps) {
 
           <FormField
             control={form.control}
-            name="recipients"
+            name="validatorAddress"
             render={({ field }) => (
               <FormItem>
-                <FormLabel>Recipients</FormLabel>
-                <FormControl>
-                  <Input placeholder="Recipient" {...field} />
-                </FormControl>
-                <FormMessage />
+                <>
+                  <FormLabel>Validators</FormLabel>
+                  <FormControl>
+                    <ValidatorSelector
+                      validators={validators.filter((validator) => {
+                        const chainId = form.watch("chainId");
+                        return chainId === ""
+                          ? true
+                          : validator.chainId === chainId;
+                      })}
+                      selectedValue={
+                        form.getValues().validatorIndex
+                          ? validators[
+                              form.getValues().validatorIndex as number
+                            ]
+                          : undefined
+                      }
+                      onSelect={(validator, index) => {
+                        form.setValue("validatorIndex", index);
+                        form.setValue("chainId", validator.chainId);
+                        form.setValue("validatorAddress", validator.address);
+                        setDecimals(validator.decimals);
+                      }}
+                      {...field}
+                    />
+                  </FormControl>
+                  <FormMessage />
+                </>
               </FormItem>
             )}
           />
